+++ eddrit/models/subreddit.py
@@ -10,10 +10,11 @@ class Subreddit:
     public_description: str | None
     over18: bool
     icon_url: str | None
+    wiki_enabled: bool
 
 
 class SubredditSortingMode(Enum):
-    POPULAR = "popular"
+    HOT = "hot"
     NEW = "new"
     RISING = "rising"
     CONTROVERSIAL = "controversial"