+++ eddrit/routes/common/exception_handlers.py
@@ -10,7 +10,6 @@ async def http_exception(request: Any, exc: Exception) -> Response:
     # If it's an HTTPException or an RedditContentUnavailableError
     # there should be a status_code
     status_code = exc.status_code if hasattr(exc, "status_code") else 500  # type: ignore
-
     # Same for the detail
     detail = exc.detail if hasattr(exc, "detail") else None  # type: ignore
     return templates.TemplateResponse(